remove intermediate functions and fixed tests
