+++ tests/fixture_plugins/demultiplex_fixtures/flow_cell_fixtures.py
@@ -118,6 +118,16 @@ def tmp_bcl_convert_flow_cell(
     )
 
 
+@pytest.fixture
+def bcl_convert_flow_cell(
+    bclconvert_flow_cell_dir_name: Path,
+    illumina_demultiplexed_runs_directory: Path,
+) -> FlowCellDirectoryData:
+    """Return a flow cell object with flow cell that is demultiplexed."""
+    path = Path(illumina_demultiplexed_runs_directory, bclconvert_flow_cell_dir_name)
+    return FlowCellDirectoryData(flow_cell_path=path)
+
+
 # Flow cell attributes
