+++ tests/fixture_plugins/demultiplex_fixtures/name_fixtures.py
@@ -1,4 +1,10 @@
+from pathlib import Path
+
 import pytest
+
+from cg.meta.demultiplex.housekeeper_storage_functions import (
+    add_and_include_sample_sheet_path_to_housekeeper,
+)
 
 
 @pytest.fixture
@@ -81,6 +87,20 @@ def bcl_convert_flow_cell_full_name() -> str:
 def novaseq_x_flow_cell_full_name() -> str:
     """Return the full name of a NovaSeqX flow cell."""
     return "20230508_LH00188_0003_A22522YLT3"
+
+
+@pytest.fixture
+def bclconvert_flow_cell_dir_name(demux_post_processing_api) -> str:
+    """Return a flow cell name that has been demultiplexed with bclconvert."""
+    flow_cell_dir_name = "230504_A00689_0804_BHY7FFDRX2"
+    flow_cell_path = Path(demux_post_processing_api.demultiplexed_runs_dir, flow_cell_dir_name)
+
+    add_and_include_sample_sheet_path_to_housekeeper(
+        flow_cell_directory=flow_cell_path,
+        flow_cell_name="HY7FFDRX2",
+        hk_api=demux_post_processing_api.hk_api,
+    )
+    return flow_cell_dir_name
 
 
 @pytest.fixture(scope="session")
