+++ tests/meta/demultiplex/conftest.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import pytest
 
 from cg.apps.housekeeper.hk import HousekeeperAPI
-from cg.constants.demultiplexing import BclConverter, DemultiplexingDirsAndFiles
+from cg.constants.demultiplexing import DemultiplexingDirsAndFiles
 from cg.meta.demultiplex.demux_post_processing import DemuxPostProcessingAPI
 from cg.meta.demultiplex.housekeeper_storage_functions import (
     add_and_include_sample_sheet_path_to_housekeeper,
@@ -279,15 +279,6 @@ def lsyncd_target_directory(lsyncd_source_directory: Path, tmp_path_factory) -> 
 
 
 @pytest.fixture
-def demux_post_processing_api(
-    demultiplex_context: CGConfig, tmp_illumina_demultiplexed_flow_cells_directory
-) -> DemuxPostProcessingAPI:
-    api = DemuxPostProcessingAPI(demultiplex_context)
-    api.demultiplexed_runs_dir = tmp_illumina_demultiplexed_flow_cells_directory
-    return api
-
-
-@pytest.fixture
 def bcl2fastq_flow_cell_dir_name(demux_post_processing_api) -> str:
     """Return a flow cell name that has been demultiplexed with bcl2fastq."""
     flow_cell_dir_name = "170407_ST-E00198_0209_BHHKVCALXX"
@@ -302,28 +293,9 @@ def bcl2fastq_flow_cell_dir_name(demux_post_processing_api) -> str:
 
 
 @pytest.fixture
-def bcl_convert_sample_id_with_non_pooled_undetermined_reads() -> str:
-    return "ACC11927A2"
-
-
-@pytest.fixture
 def bcl2fastq_non_pooled_sample_read_count() -> int:
     """Based on the data in 170407_ST-E00198_0209_BHHKVCALXX, the sum of all reads - mapped and undetermined."""
     return 8000000
-
-
-@pytest.fixture
-def bclconvert_flow_cell_dir_name(demux_post_processing_api) -> str:
-    """Return a flow cell name that has been demultiplexed with bclconvert."""
-    flow_cell_dir_name = "230504_A00689_0804_BHY7FFDRX2"
-    flow_cell_path = Path(demux_post_processing_api.demultiplexed_runs_dir, flow_cell_dir_name)
-
-    add_and_include_sample_sheet_path_to_housekeeper(
-        flow_cell_directory=flow_cell_path,
-        flow_cell_name="HY7FFDRX2",
-        hk_api=demux_post_processing_api.hk_api,
-    )
-    return flow_cell_dir_name
 
 
 @pytest.fixture
